requirements.txt
async-generator==1.10
transformers
scikit-learn
datasets
flask
attrs==22.2.0
beautifulsoup4==4.11.2
blinker==1.6.2
boto3
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
DateTime==5.0
dateparser
distlib==0.3.6
filelock==3.8.0
feedparser
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.2
outcome==1.2.0
pandas==1.5.3
platformdirs==2.5.4
PySocks==1.7.1
python-dateutil==2.8.2
pytrends
pygithub
pytz==2022.7.1
python-dotenv
rapidfuzz==2.13.7
requests==2.28.2
s3transfer==0.6.0
selenium==4.8.0
simplejson==3.18.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
virtualenv==20.17.0
Werkzeug==2.3.4
wsproto==1.2.0
zope.interface==5.5.2